pyproject.toml
[tool.poetry]
name = "exareme2"
version = "0.1.0"
description = "Core engine for the Medical Informatics Platform"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "~3.8"
hypercorn = "^0.14.4"
Quart = "~0.18"
celery = "~5.3.1"
asgiref = "~3.7"
pymonetdb = "1.6.3" # Must be updated together with mipdb Dockerfile
astor = "~0.8"
numpy = "1.24.1" # Must be updated together with monetdb Dockerfile
scipy = "1.10.0" # Must be updated together with monetdb Dockerfile
pandas = "1.5.2" # Must be updated together with monetdb Dockerfile
scikit-learn = "1.2.0"  # Must be updated together with monetdb Dockerfile
dnspython = "~2.3.0" # TODO upgrade this when: https://github.com/rthalley/dnspython/issues/957
pydantic = "~1.10"
envtoml = "~0.1"
requests = "~2.31"
eventlet = "~0.33"
patsy = "~0.5.3"

[tool.poetry.dev-dependencies]
pytest = "~7.4"
pytest-xdist = "~3.3"
pytest-asyncio="~0.21"
pytest-mock="~3.11"
pytest-assume="~2.4"
ipython = "~8.12"
pdbpp = "~0.10"
invoke = "~2.2"
black = "23.7.0" # Must be upgraded together with github actions
isort = "5.12.0" # Must be upgraded together with github actions
pre-commit = "~3.3"
termcolor = "~2.3"
coverage = {extras = ["toml"], version = "~7.2"}
pytest-cov = "~4.1"
pylint = "~2.17"
rope = "~1.9"
devtools="~0.11"
docker = "~6.1"
tqdm = "~4.65"
SQLAlchemy = "~1.3.0" # TODO https://team-1617704806227.atlassian.net/browse/MIP-740
statsmodels = "0.13.2" # Must be updated together with monetdb Dockerfile
psutil = "~5.9"
pytest-subtests = "~0.11"
freezegun = "~1.2"
hypothesis = "~6.81"
pytest-rerunfailures = "~12.0"


### GRPC
grpcio-tools = "~1.56"
grpcio-status = "~1.56"
googleapis-common-protos = "~1.59"

[tool.poetry.group.dev.dependencies]
mipdb = "2.1.0"  # Must be updated together with mipdb Dockerfile

[tool.pytest.ini_options]
markers = [
    "database: these tests depend on an external dockerized MonetDB service running (deselect with '-m \"not database\"')",
    "slow: marks tests as slow, mostly integration tests (deselect with '-m \"not slow\"')",
    "very_slow: marks tests as very_slow, integration tests with temporary container dependencies (deselect with '-m \"not very_slow\"')",
    "smpc: marks the tests validating smpc functionality (deselect with '-m \"not smpc\"')",
    "smpc_cluster: marks the tests that need smpc cluster deployment (deselect with '-m \"not smpc_cluster\"')",
]
filterwarnings = ["ignore::DeprecationWarning"]
norecursedirs = ["tests/testcase_generators"]

[tool.coverage.run]
source_pkgs = ['exareme2']
dynamic_context = 'test_function'

[tool.coverage.report]
exclude_lines = ['def __repr__', 'raise NotImplementedError', 'pragma: no cover']
show_missing = true

[tool.coverage.html]
show_contexts = true

[build-system]
requires = ["poetry-core>=1.4"]
build-backend = "poetry.core.masonry.api"
